Add quotes to variable values and improve readability
The main problem was unquoted value in "ftp_protocol_cmd=set ftps:initial-prot P" line in upload scripts, which contained spaces and interpreted as a command line.

Additionaly, `backquoted` call output to variable syntax is simpler than "$(...)" and highlighted with a nice different color in my text editor.

Whether to quote hostnames is questionable, so it was done only where it improved readability.