a.py: pick Zstd level from count of repeated '9' flags
Results for Zstd level 17 and 20 differ enough to try both.
And level 3 may be used as a substitute for "store" method.

Still, LZMA2 compresses better than Zstd level 22.
And WinRAR with "store" method is better for fast backups of currently used files (like a profile folder of a running web-browser), than 7-Zip with any Zstd.